backend/requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
python-dotenv
requests
python-multipart
pydantic-settings
sqlalchemy
alembic
reportlab
python-pptx